backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.7
anyio==3.7.1
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.131
botocore==1.34.131
cachetools==5.3.3
certifi==2024.2.2
chainlit==1.1.306
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
construct==2.5.3
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.20.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
faiss-cpu==1.8.0
fastapi==0.110.3
fastapi-cli==0.0.4
filelock==3.14.0
filetype==1.2.0
FLAML==2.1.2
Flask==3.0.3
fonttools==4.53.0
free_proxy==1.1.1
frozenlist==1.4.1
fsspec==2024.3.1
google==3.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
google-search-results==2.4.2
googleapis-common-protos==1.63.0
graphviz==0.20.3
greenlet==3.0.3
groq==0.8.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
haystack==0.42
haystack-ai==2.2.4
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.6
langchain-anthropic==0.1.11
langchain-aws==0.1.3
langchain-community==0.0.36
langchain-core==0.2.10
langchain-experimental==0.0.57
langchain-google-genai==1.0.3
langchain-groq==0.1.3
langchain-openai==0.1.6
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langserve==0.2.2
langsmith==0.1.81
Lazify==0.4.0
lazy-imports==0.3.1
literalai==0.0.607
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
minify_html==0.15.0
monotonic==1.6
more-itertools==10.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.25.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pefile==2023.2.7
pillow==10.4.0
pinecone-client==3.2.2
platformdirs==4.2.2
playwright==1.43.0
pluggy==1.5.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.32
pydantic==2.7.1
pydantic-settings==2.3.4
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproject-toml==0.0.10
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-ptrace==0.9.9
python-socketio==5.11.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.1
scrapegraphai==1.7.4
semchunk==1.0.1
setuptools==68.2.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sse-starlette==2.1.2
starlette==0.37.2
syncer==2.0.3
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
undetected-playwright==0.3.0
uptrace==1.24.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.25.0
virtualenv==20.26.3
watchfiles==0.20.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
